fix: update send-email action and appointment page for messageEn
Correct return type in send-email function and page.tsx to support messageEn.
